feat:selection write, enter를 Edit으로 통합
